user_info_check.html user_register.html 추가

+++ Backend/src/main/java/kr/ac/kumoh/ce/s20210353/Child/Safety/Tag/service/NfcTagService.java
@@ -27,7 +27,11 @@ public class NfcTagService {
         return repository.save(tag);
     }
 
-    public void deleteById(NfcTag.NfcTagId id) {
-        repository.deleteById(id);
+    public boolean deleteById(NfcTag.NfcTagId id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
